+++ packages/ui-extensions/src/surfaces/checkout/components/Image/Image.doc.ts
@@ -22,7 +22,7 @@ const data: ReferenceEntityTemplateSchema = {
       tabs: [
         {
           code: './examples/basic-image.example.html',
-          language: '',
+          language: 'html',
         },
       ],
     },