+++ packages/ui-extensions/src/surfaces/customer-account/components/ImageGroup/ImageGroup.doc.ts
@@ -3,7 +3,7 @@ import type {ReferenceEntityTemplateSchema} from '@shopify/generate-docs';
 const data: ReferenceEntityTemplateSchema = {
   name: 'ImageGroup',
   description:
-    'Display up to 4 images in a grid or stacked layout. For example, images of products in a wishlist or subscription. When there are more than 4 images, the component indicates how many more images are not displayed.',
+    'Display up to 4 images in a grid or stacked layout. The images are displayed as a grid when used within a [Section](/docs/api/customer-account-ui-extensions/polaris-web-components/structure/section) component. For example, images of products in a wishlist or subscription. When there are more than 4 images, the component indicates how many more images are not displayed.',
   thumbnail: 'imagegroup-thumbnail.png',
   requires: '',
   isVisualComponent: true,
@@ -18,16 +18,15 @@ const data: ReferenceEntityTemplateSchema = {
   category: 'Polaris web components',
   subCategory: 'Media',
   defaultExample: {
-    image: 'imagegroup-preview.png',
+    image: 'imagegroup-default.png',
     altText:
       'An example of the ImageGroup component shows a group of four images of plants, arranged in a 2x2 grid.',
     codeblock: {
-      title: 'Basic ImageGroup',
+      title: 'Code',
       tabs: [
         {
-          title: 'Preact',
-          code: './examples/basic-ImageGroup-preact.example.tsx',
-          language: 'tsx',
+          code: './examples/basic-ImageGroup.example.html',
+          language: 'jsx',
         },
       ],
     },