+++ packages/ui-extensions/src/surfaces/point-of-sale/components/Badge/Badge.doc.ts
@@ -3,7 +3,7 @@ import {ReferenceEntityTemplateSchema} from '@shopify/generate-docs';
 const data: ReferenceEntityTemplateSchema = {
   name: 'Badge',
   description:
-    "Use `s-badge` to inform merchants of the status of an item or action that's been taken. Badges display text with visual styling to communicate status information.",
+    'Informs merchants about the status of an object or indicates that an action has been completed. Badges display text with visual styling to communicate status information.',
   thumbnail: 'badge-thumbnail.png',
   isVisualComponent: true,
   type: '',
@@ -28,6 +28,33 @@ const data: ReferenceEntityTemplateSchema = {
       ],
     },
   },
+  subSections: [
+    {
+      type: 'Generic',
+      anchorLink: 'guidelines',
+      title: 'Guidelines',
+      sectionContent: `
+- Badges should be positioned as close as possible to the item they’re related to.
+`,
+    },
+    {
+      type: 'Generic',
+      anchorLink: 'content-guidelines',
+      title: 'Content guidelines',
+      sectionContent: `
+- Be concise. Use a single word to describe the status of an item.
+- Only use two or three words if you need to describe a complex state, for example "partially fulfilled".
+
+✅ fulfilled
+❌ order fulfilled
+
+Statuses should ideally be written as adjectives:
+
+✅ unpaid
+❌ payment not received
+      `,
+    },
+  ],
   related: [],
 };