+++ packages/ui-extensions/src/surfaces/point-of-sale/components/PosBlock/PosBlock.doc.ts
@@ -3,7 +3,7 @@ import {ReferenceEntityTemplateSchema} from '@shopify/generate-docs';
 const data: ReferenceEntityTemplateSchema = {
   name: 'PosBlock',
   description:
-    'Use `s-pos-block` to create a surface on the specified block extension target, serving as a container to place content with an action button.',
+    'Creates a surface on the specified block extension target, serving as a container to place content with an action button.',
   thumbnail: 'pos-block-thumbnail.png',
   isVisualComponent: true,
   type: '',